refactor(ui): start view-model.ts split — move shared types + terminology to view-model/types.ts
First slice of the 642-LOC view-model.ts decomposition. Hoist the
SYNC_TERMINOLOGY const and all exported type definitions (ActorLike,
PeerLike, DiscoveredDeviceLike, Ui* interfaces, state unions,
VisiblePeopleResult, SyncPeerStatusLike) into a dedicated types
module. The view-model.ts barrel re-exports everything so existing
call sites across the sync tabs and state.ts stay on the same import
path.
