+++ packages/ui/src/tabs/sync/view-model.ts
@@ -1,122 +1,36 @@
 /* Derived sync view-model helpers. Keep these pure so UX logic is testable. */
 
-export const SYNC_TERMINOLOGY = {
-	actor: "person",
-	actors: "people",
-	actorAssignment: "person assignment",
-	localActor: "you",
-	peer: "device",
-	peers: "devices",
-	pairedLocally: "Connected on this device",
-	discovered: "Seen on team",
-	conflicts: "Needs review",
-} as const;
+import type {
+	ActorLike,
+	DiscoveredDeviceLike,
+	PeerLike,
+	UiCoordinatorApprovalSummary,
+	UiDuplicatePersonCandidate,
+	UiPeerTrustSummary,
+	UiSyncAttentionItem,
+	UiSyncRunResponse,
+	UiSyncStatus,
+	UiSyncViewModel,
+	VisiblePeopleResult,
+} from "./view-model/types";
 
-export type UiSyncStatus = "connected" | "available" | "needs-repair" | "offline" | "waiting";
-export type UiTrustState =
-	| "available"
-	| "trusted-by-you"
-	| "mutual-trust"
-	| "needs-repairing"
-	| "needs-review"
-	| "offline";
-export type UiCoordinatorApprovalState =
-	| "none"
-	| "needs-your-approval"
-	| "waiting-for-other-device";
-
-interface SyncPeerStatusLike {
-	peer_state?: string;
-	sync_status?: string;
-	ping_status?: string;
-	fresh?: boolean;
-}
-
-export interface UiSyncRunItem {
-	peer_device_id: string;
-	ok: boolean;
-	error?: string;
-	address?: string;
-	opsIn: number;
-	opsOut: number;
-	addressErrors: Array<{ address: string; error: string }>;
-}
-
-export interface UiSyncRunResponse {
-	items: UiSyncRunItem[];
-}
-
-export interface UiSyncAttentionItem {
-	id: string;
-	kind: "possible-duplicate-person" | "device-needs-repair" | "review-team-device" | "name-device";
-	priority: number;
-	title: string;
-	summary: string;
-	actionLabel: string;
-	deviceId?: string;
-	actorIds?: string[];
-}
-
-export interface UiDuplicatePersonCandidate {
-	displayName: string;
-	actorIds: string[];
-	includesLocal: boolean;
-}
-
-export interface UiSyncViewModel {
-	summary: {
-		connectedDeviceCount: number;
-		seenOnTeamCount: number;
-		offlineTeamDeviceCount: number;
-	};
-	duplicatePeople: UiDuplicatePersonCandidate[];
-	attentionItems: UiSyncAttentionItem[];
-}
-
-export interface UiPeerTrustSummary {
-	state: UiTrustState;
-	badgeLabel: string;
-	description: string;
-	isWarning: boolean;
-}
-
-export interface VisiblePeopleResult {
-	visibleActors: ActorLike[];
-	hiddenLocalDuplicateCount: number;
-}
-
-export interface ActorLike {
-	actor_id?: string;
-	display_name?: string;
-	is_local?: boolean;
-}
-
-export interface PeerLike {
-	peer_device_id?: string;
-	name?: string;
-	has_error?: boolean;
-	last_error?: string;
-	fingerprint?: string;
-	actor_id?: string;
-	status?: SyncPeerStatusLike;
-}
-
-export interface DiscoveredDeviceLike {
-	device_id?: string;
-	display_name?: string;
-	stale?: boolean;
-	fingerprint?: string;
-	groups?: string[];
-	needs_local_approval?: boolean;
-	waiting_for_peer_approval?: boolean;
-}
-
-export interface UiCoordinatorApprovalSummary {
-	state: UiCoordinatorApprovalState;
-	badgeLabel: string | null;
-	description: string | null;
-	actionLabel: string | null;
-}
+export {
+	type ActorLike,
+	type DiscoveredDeviceLike,
+	type PeerLike,
+	SYNC_TERMINOLOGY,
+	type UiCoordinatorApprovalState,
+	type UiCoordinatorApprovalSummary,
+	type UiDuplicatePersonCandidate,
+	type UiPeerTrustSummary,
+	type UiSyncAttentionItem,
+	type UiSyncRunItem,
+	type UiSyncRunResponse,
+	type UiSyncStatus,
+	type UiSyncViewModel,
+	type UiTrustState,
+	type VisiblePeopleResult,
+} from "./view-model/types";
 
 interface MergedDevice {
 	deviceId: string;
